+++ src/imio/smartweb/core/webcomponents/src/components/News/NewsContent/NewsContent.jsx
@@ -244,7 +244,7 @@ const ContactContent = ({ queryUrl, onChange }) => {
 						<div class="spotlight-group flexbin r-content-gallery">
 							{gallery.map((image,i) => (
 								<a key={i} class="spotlight" href={image.image_full_scale} >
-									<img src={image.image_preview_scale} />
+									<img src={image.image_preview_scale} alt="" />
 								</a>
 							))}
 						</div>